correción en devoluciones y search

+++ app/User.php
@@ -16,53 +16,59 @@ class User extends Authenticatable
 
         parent::boot();
 
-       /* static::created(function(User $user) {
-			
-			$slug = \Str::slug($user->nombres. " " . $user->apellidos);
-			
-			$count = static::whereRaw("slug RLIKE '^{$slug}(-[0-9]+)?$'")->count();
-			
+        static::creating(function(User $user) {
+      
+           $slug = \Str::slug($user->nombres. " " . $user->apellidos);
+                
+            $count = static::whereRaw("slug RLIKE '^{$slug}(-[0-9]+)?$'")->count();
+            
             $user->slug = $count ? "{$slug}-{$count}" : $slug;
             
-            if (request()->file('imagen')) {
+        });
 
-                $imagen = request()->file('imagen');
-                $nombre = time().'_'.$imagen->getClientOriginalName();
-                //$imageName = \Str::random(20) . '.jpg';
-                $image = Image::make($imagen)->encode('jpg', 75);
-                $image->resize(128, 128, function ($constraint){
-                    $constraint->upsize();
-                });
+        static::created(function(User $user) {
+
+            try {
                 
-                Storage::disk('dropbox')->put("users/$nombre", $image->stream()->__toString());
-                $dropbox = Storage::disk('dropbox')->getDriver()->getAdapter()->getClient();
-                $response = $dropbox->createSharedLinkWithSettings("users/$nombre", ["requested_visibility" => "public"]);
-                $path = str_replace('dl=0', 'raw=1', $response['url']);
-                $imageName = $response['name'];
+                // $slug = \Str::slug($user->nombres. " " . $user->apellidos);
+                
+                // $count = static::whereRaw("slug RLIKE '^{$slug}(-[0-9]+)?$'")->count();
+                
+                // $user->slug = $count ? "{$slug}-{$count}" : $slug;
 
-                $img = new Imagene();
-                $img->nombre = $imageName;
-                $img->url = $path;
-                $img->imageable_type = 'App\User';
-                $img->imageable_id = $user->id;
-
-                $img->save();
-
-                Cliente::create([
-                    'user_id' => $user->id,
-                ]);
-            
-            }
-			
-		});*/
-
-        //implementar con dropbox
-		static::saved(function(User $user) {
-			
-			if( ! \App::runningInConsole() ) {
-
-				if (request()->file('imagen')) {
+                
+                // if (request()->file('imagen')) {
     
+                //     $imagen = request()->file('imagen');
+                //     $nombre = time().'_'.$imagen->getClientOriginalName();
+                //     //$imageName = \Str::random(20) . '.jpg';
+                //     $image = Image::make($imagen)->encode('jpg', 75);
+                //     $image->resize(128, 128, function ($constraint){
+                //         $constraint->upsize();
+                //     });
+                    
+                //     Storage::disk('dropbox')->put("users/$nombre", $image->stream()->__toString());
+                //     $dropbox = Storage::disk('dropbox')->getDriver()->getAdapter()->getClient();
+                //     $response = $dropbox->createSharedLinkWithSettings("users/$nombre", ["requested_visibility" => "public"]);
+                //     $path = str_replace('dl=0', 'raw=1', $response['url']);
+                //     $imageName = $response['name'];
+    
+                //     $img = new Imagene();
+                //     $img->nombre = $imageName;
+                //     $img->url = $path;
+                //     $img->imageable_type = 'App\User';
+                //     $img->imageable_id = $user->id;
+    
+                //     $img->save();
+    
+                    // Cliente::create([
+                    //     'user_id' => $user->id,
+                    // ]);
+                
+                // }
+    
+                if (request()->file('imagen')) {
+        
                     $imagen = request()->file('imagen');
                     $nombre = time().'_'.$imagen->getClientOriginalName();
                     //$image = Image::make($imagen)->encode('jpg', 75);
@@ -79,30 +85,79 @@ class User extends Authenticatable
     
                     $img = new Imagene();
                     //$img->nombre = $imageName;
-                    $img->nombre = $nombre;
+                    // $img->nombre = $nombre;
                     $img->url = $path;
                     $img->imageable_type = 'App\User';
                     $img->imageable_id = $user->id;
-
+    
                     $img->save();
-
-                    if ($user->cliente) {
-                        $imagen = Imagene::where('imageable_type','App\User')
-                        ->where('imageable_id', auth()->user()->id)->first();
-
-                        if ($imagen != '') {
-                            //$delete = $this->dropbox->delete($imagen->nombre);
-                            Storage::disk('public')->delete($imagen->url);
-                            $imagen->delete();
-                        }
-
-                    } else {
-                        Cliente::create([
-                            'user_id' => $user->id,
-                        ]);
-                    }
-                   
                 }
+    
+                Cliente::create([
+                    'user_id' => $user->id,
+                ]);
+                
+            } catch (\Exception $e) {
+               
+            }
+			
+		});
+
+        //implementar con dropbox
+		static::saving(function(User $user) {
+			
+			if( ! \App::runningInConsole() ) {
+
+                try {
+                   
+                    if (request()->file('imagen')) {
+        
+                        $imagen = request()->file('imagen');
+                        $nombre = time().'_'.$imagen->getClientOriginalName();
+                        //$image = Image::make($imagen)->encode('jpg', 75);
+                        //$image->resize(128, 128, function ($constraint){
+                            //$constraint->upsize();
+                        //});
+                        
+                        //Storage::disk('dropbox')->put("users/$nombre", $image->stream()->__toString());
+                        //$dropbox = Storage::disk('dropbox')->getDriver()->getAdapter()->getClient();
+                        //$response = $dropbox->createSharedLinkWithSettings("users/$nombre", ["requested_visibility" => "public"]);
+                        //$path = str_replace('dl=0', 'raw=1', $response['url']);
+                        //$imageName = $response['name'];
+                        $path = Storage::disk('public')->putFileAs("imagenes/users/" . $user->id, $imagen, $nombre);
+        
+                        $img = new Imagene();
+                        //$img->nombre = $imageName;
+                        // $img->nombre = $nombre;
+                        $img->url = $path;
+                        $img->imageable_type = 'App\User';
+                        $img->imageable_id = $user->id;
+    
+                        $img->save();
+    
+                        if ($user->cliente) {
+                            $imagen = Imagene::where('imageable_type','App\User')
+                            ->where('imageable_id', auth()->user()->id)->first();
+    
+                            if ($imagen != '') {
+                                //$delete = $this->dropbox->delete($imagen->nombre);
+                                Storage::disk('public')->delete($imagen->url);
+                                $imagen->delete();
+                            }
+    
+                        } 
+                        // else {
+                        //     Cliente::create([
+                        //         'user_id' => $user->id,
+                        //     ]);
+    
+                        // }
+                       
+                    }
+                } catch (\Exception $e) {
+                    //throw $th;
+                }
+
 			}
 		});
     }
@@ -115,7 +170,8 @@ class User extends Authenticatable
      * @var array
      */
     protected $fillable = [
-    	'identificacion','nombres', 'apellidos', 'direccion', 'telefono', 'email', 'username','password','imagen', 'slug'
+    	'identificacion','nombres', 'apellidos', 'direccion', 'telefono', 'email',
+         'username','password','imagen', 'slug', 'departamento', 'municipio'
     ];
 
     /**
